requirements.txt
pyqt5>=5.15.10
loguru>=0.7.2
onnx>=1.16.0
onnxruntime>=1.17.3
setuptools>=69.5.1
pyyaml>=6.0.1